simplify - no need for traits
